+++ infobip-openapi-mcp-core/src/main/java/com/infobip/openapi/mcp/openapi/schema/Spec.java
@@ -1,0 +1,13 @@
+package com.infobip.openapi.mcp.openapi.schema;
+
+import java.util.Set;
+
+class Spec {
+    private Spec() {}
+
+    static Set<String> SUPPORTED_PARAMETER_TYPES = Set.of(
+            DecomposedRequestData.ParametersByType.QUERY,
+            DecomposedRequestData.ParametersByType.PATH,
+            DecomposedRequestData.ParametersByType.HEADER,
+            DecomposedRequestData.ParametersByType.COOKIE);
+}